+++ doc/source/releasenotes.rst
@@ -4,6 +4,10 @@
 Release notes
 =============
 
+1.0.2
+======
+
+
 1.0.1
 ======
 * Added the ``ignore_sizes`` keyword to :py:func:`~clease.plot_post_process.plot_eci`